Bump to Magnesium dependencies
https://docs.opendaylight.org/projects/integration-distribution/en/latest/platform-versions.html
- upgrade pom to use ODL Magnesium target dev dependecies

https://wiki.opendaylight.org/view/Magnesium_platform_upgrade
- force jsr305 version
- disable SpotBugs enforcement in artifacts when needed
- UintXX yangtools migration

https://git.opendaylight.org/gerrit/c/mdsal/+/83960
- migrate class BindingStreamEventWriter
      from org.opendaylight.yangtools.yang.binding package
      to   org.opendaylight.mdsal.binding.dom.codec.api

- unforce maven-bundle-plugin version in api/pom.xml
  and ordmodels/device/pom.xml build sections
- unfoce iana-afn-safi version in ordmodels/device/pom.xml dependencies

TODO:
for next changes to come in the release:
- migrate Junit to yangtools UintXX
- propose fixes to reactivate SpotBugs enforcement
  and get rid of "deprecated marked for removal"

JIRA: TRNSPRTPCE-198

+++ pce/src/main/java/org/opendaylight/transportpce/pce/networkanalyzer/PceResult.java
@@ -15,6 +15,7 @@ import org.opendaylight.transportpce.common.ResponseCodes;
 import org.opendaylight.yang.gen.v1.http.org.openroadm.service.format.rev190531.ServiceFormat;
 import org.opendaylight.yang.gen.v1.http.org.transportpce.b.c._interface.pathdescription.rev171017.path.description.AToZDirection;
 import org.opendaylight.yang.gen.v1.http.org.transportpce.b.c._interface.pathdescription.rev171017.path.description.ZToADirection;
+import org.opendaylight.yangtools.yang.common.Uint16;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -25,8 +26,8 @@ public class PceResult {
     private boolean calcStatus = false;
     private String responseCode = ResponseCodes.RESPONSE_FAILED;
     private long resultWavelength = -1;
-    private Map<String, Integer> resultTribPort;
-    private Map<String, List<Integer>> resultTribSlot;
+    private Map<String, Uint16> resultTribPort;
+    private Map<String, List<Uint16>> resultTribSlot;
     private Integer resultTribSlotNb = -1;
     private String serviceType = "";
 
@@ -137,19 +138,19 @@ public class PceResult {
         this.calcMessage = calcMessage;
     }
 
-    public Map<String, Integer> getResultTribPort() {
+    public Map<String, Uint16> getResultTribPort() {
         return resultTribPort;
     }
 
-    public void setResultTribPort(Map<String, Integer> resultTribPort) {
+    public void setResultTribPort(Map<String, Uint16> resultTribPort) {
         this.resultTribPort = resultTribPort;
     }
 
-    public Map<String, List<Integer>> getResultTribSlot() {
+    public Map<String, List<Uint16>> getResultTribSlot() {
         return resultTribSlot;
     }
 
-    public void setResultTribSlot(Map<String, List<Integer>> resultTribSlot) {
+    public void setResultTribSlot(Map<String, List<Uint16>> resultTribSlot) {
         this.resultTribSlot = resultTribSlot;
     }
